+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `hybrid` map type & removed `raster`, `map3D` map type
 * Removed the `enable3DMode` method from `MFMapViewController`
+* Remove `MFMapView.onModeChange` callback
 
 ## 2.6.0